# Slight performance improvement of "reinterpret" function