Changelog category (leave one):
- Performance Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Slight performance improvement of `reinterpret` function.


Detailed description / Documentation draft:

Was:

```
localhost:9000, queries 24, QPS: 2.068, RPS: 413693233.016, MiB/s: 3156.229, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.452 sec.
10.000%         0.458 sec.
20.000%         0.460 sec.
30.000%         0.463 sec.
40.000%         0.470 sec.
50.000%         0.480 sec.
60.000%         0.484 sec.
70.000%         0.495 sec.
80.000%         0.513 sec.
90.000%         0.521 sec.
95.000%         0.527 sec.
99.000%         0.536 sec.
99.900%         0.536 sec.
99.990%         0.536 sec.
```

Now:

```
localhost:9000, queries 34, QPS: 2.122, RPS: 424415529.072, MiB/s: 3238.034, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.405 sec.
10.000%         0.435 sec.
20.000%         0.450 sec.
30.000%         0.455 sec.
40.000%         0.464 sec.
50.000%         0.471 sec.
60.000%         0.478 sec.
70.000%         0.488 sec.
80.000%         0.495 sec.
90.000%         0.505 sec.
95.000%         0.519 sec.
99.000%         0.531 sec.
99.900%         0.531 sec.
99.990%         0.531 sec.
```